requirements.txt
pandas>=1.5.0
google-generativeai>=0.3.0
plotly>=5.13.0
dash>=2.9.0
dash-bootstrap-components>=1.4.0
python-dotenv>=0.19.0 